+++ packages/eslint-config-airbnb-base/rules/variables.js
@@ -12,11 +12,11 @@ module.exports = {
     'no-delete-var': 'error',
 
     // disallow labels that share a name with a variable
-    // http://eslint.org/docs/rules/no-label-var
+    // https://eslint.org/docs/rules/no-label-var
     'no-label-var': 'error',
 
     // disallow specific globals
-    'no-restricted-globals': ['error'].concat(restrictedGlobals),
+    'no-restricted-globals': ['error', 'isFinite', 'isNaN'].concat(restrictedGlobals),
 
     // disallow declaration of variables already declared in the outer scope
     'no-shadow': 'error',
@@ -31,7 +31,7 @@ module.exports = {
     'no-undef-init': 'error',
 
     // disallow use of undefined variable
-    // http://eslint.org/docs/rules/no-undefined
+    // https://eslint.org/docs/rules/no-undefined
     // TODO: enable?
     'no-undefined': 'off',