+++ packages/eslint-config-airbnb-base/test/test-base.js
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import test from 'tape';
 
-import index from '../';
+import index from '..';
 
 const files = { ...{ index } }; // object spread is to test parsing
 
@@ -11,17 +11,17 @@ fs.readdirSync(path.join(__dirname, '../rules')).forEach((name) => {
   files[name] = require(`../rules/${name}`); // eslint-disable-line global-require
 });
 
-Object.keys(files).forEach((
+Object.keys(files).forEach(( // eslint-disable-line function-paren-newline
   name, // trailing function comma is to test parsing
-) => {
+) => { // eslint-disable-line function-paren-newline
   const config = files[name];
 
   test(`${name}: does not reference react`, (t) => {
     t.plan(2);
 
     // scan plugins for react and fail if it is found
-    const hasReactPlugin = Object.prototype.hasOwnProperty.call(config, 'plugins') &&
-      config.plugins.indexOf('react') !== -1;
+    const hasReactPlugin = Object.prototype.hasOwnProperty.call(config, 'plugins')
+      && config.plugins.indexOf('react') !== -1;
     t.notOk(hasReactPlugin, 'there is no react plugin');
 
     // scan rules for react/ and fail if any exist